[fix] get space from url
