--- src/components/settings/SpaceSettings/SpaceSettingsInvite.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function SpaceSettingsInvite() {
-  return (
-    <div>
-      <p>Invitations</p>
-    </div>
-  );
-}
-
-export default SpaceSettingsInvite;
