+++ src/utils/getSpaceID.js
@@ -1,5 +1,7 @@
-const spaceId = window.location.href.substring(
-  window.location.href.lastIndexOf('-') + 1
-);
-
+/**
+ * get id space from url
+ */
+const spaceId = window.location.href
+  .split('/')[4]
+  .substring(window.location.href.split('/')[4].lastIndexOf('-') + 1);
 export default spaceId;
